+++ internal/dto/memeception.go
@@ -1,27 +1,27 @@
 package dto
 
 type MemeceptionCommon struct {
-	StartAt         uint   `json:"startAt"`
-	Status          uint   `json:"status"`
+	StartAt         uint64 `json:"startAt"`
+	Status          uint64 `json:"status"`
 	Ama             bool   `json:"ama"`
 	ContractAddress string `json:"contractAddress"`
 	TargetETH       string `json:"targetETH"`
 	CollectedETH    string `json:"collectedETH"`
 	Enabled         bool   `json:"enabled"`
-	MemeID          uint   `json:"memeID"`
-	UpdatedAtEpoch  uint   `json:"updatedAtEpoch"`
+	MemeID          uint64 `json:"memeID"`
+	UpdatedAtEpoch  uint64 `json:"updatedAtEpoch"`
 }
 
 type Memeception struct {
-	StartAt         uint       `json:"startAt"`
-	Status          uint       `json:"status"`
+	StartAt         uint64     `json:"startAt"`
+	Status          uint64     `json:"status"`
 	Ama             bool       `json:"ama"`
 	ContractAddress string     `json:"contractAddress"`
 	TargetETH       string     `json:"targetETH"`
 	CollectedETH    string     `json:"collectedETH"`
 	Enabled         bool       `json:"enabled"`
-	MemeID          uint       `json:"memeID"`
-	UpdatedAtEpoch  uint       `json:"updatedAtEpoch"`
+	MemeID          uint64     `json:"memeID"`
+	UpdatedAtEpoch  uint64     `json:"updatedAtEpoch"`
 	Meme            MemeCommon `json:"meme"`
 }
